Add files via upload
This patch adds documentation and slight modifications to the main program. This should have a better consistency when calling any User object. This patch basically allows the program to have many users without any problem which can not be said to the first one where checking and savings account were improperly syncing with the wrong user